+++ app/assets/javascripts/binda/components/form_item_repeater.js
@@ -2,6 +2,7 @@
 /// FORM ITEM
 ///- - - - - - - - - - - - - - - - - - - -
 
+import { custom_fileupload } from './fileupload_custom_script'
 import { _FormItemEditor } from './form_item_editor'
 
 class FormItemRepeater {
@@ -33,6 +34,8 @@ class FormItemRepeater {
 		{
 			// Stop default behaviour
 			event.preventDefault()
+
+			if ( !confirm("Are you sure you want do delete it?") ) return
 
 			$.ajax({
 				url: $( this ).attr('href'),
@@ -69,5 +72,6 @@ function addNewItem( event )
 		var editor_id = $list.find('textarea').last('textarea').attr('id')
 		tinyMCE.EditorManager.execCommand('mceAddEditor',true, editor_id);
 		_FormItemEditor.resize()
+		custom_fileupload( $list.find('.fileupload').last('.fileupload').get(0) );
 	})
 }
